cmapPy/pandasGEXpress/parse_gctx.py: when parsing data match datatype, don't force to float32

+++ cmapPy/pandasGEXpress/parse_gctx.py
@@ -372,7 +372,7 @@ def parse_data_df(data_dset, ridx, cidx, row_meta, col_meta):
     total_rows = len(row_meta.index)
     total_cols = len(col_meta.index)
     if len(ridx) == total_rows and len(cidx) == total_cols:  # no subset
-        data_array = np.empty(data_dset.shape, dtype=np.float32)
+        data_array = np.empty(data_dset.shape, dtype=data_dset.dtype)
         data_dset.read_direct(data_array)
         data_array = data_array.transpose()
     else:
@@ -383,10 +383,10 @@ def parse_data_df(data_dset, ridx, cidx, row_meta, col_meta):
         col_first_count = total_rows * len(cidx)
 
         if row_first_count < col_first_count:
-            first_subset = data_dset[:, ridx].astype(np.float32)
+            first_subset = data_dset[:, ridx].astype(data_dset.dtype)
             data_array = first_subset[cidx, :].transpose()
         else:
-            first_subset = data_dset[cidx, :].astype(np.float32)
+            first_subset = data_dset[cidx, :].astype(data_dset.dtype)
             data_array = first_subset[:, ridx].transpose()
 
     # make DataFrame instance
